refactor: 美容室ページのロジックをリファクタリング, feat: 飲料水ページに食事メニューの追加機能を実装
