+++ src/components/restaurant/main/applicationSet/AccountSettings.tsx
@@ -6,9 +6,9 @@ import { ListBtn } from "../../../table/Table";
 function AccountSettings() {
   // 계좌정보 데이터
   const [accountData, setAccountData] = useState<AccountType>({
-    account: "계좌번호",
-    bank_name: "은행",
-    name: "예금자",
+    account: "",
+    bank_name: "",
+    name: "",
   });
   // 수정 상태
   const [onModify, setOnModify] = useState(false);
@@ -111,7 +111,7 @@ function AccountSettings() {
         <div className="flex items-start ">
           <div className="flex flex-col flex-1 gap-1 p-4 font-semibold text-lg">
             <div className="flex gap-2">
-              <div>은행 : </div>
+              <div>은행 명 : </div>
               <div>{accountData.bank_name}</div>
             </div>
             <div className="flex gap-2">
@@ -119,7 +119,7 @@ function AccountSettings() {
               <div>{accountData.account}</div>
             </div>
             <div className="flex gap-2">
-              <div>예금주 : </div>
+              <div>예금자 명 : </div>
               <div>{accountData.name}</div>
             </div>
           </div>
