--- src/pages/salon/Salon.tsx
@@ -1,12 +1,0 @@
-import MainContainer from "../../components/MainContainer";
-import ReservationList from "../../components/salon/ReservationList";
-
-function Salon() {
-  return (
-    <MainContainer>
-      <ReservationList />
-    </MainContainer>
-  );
-}
-
-export default Salon;
